Fully storing the configs in the store state, removing duplicity
