+++ config/dev.exs
@@ -1,0 +1,1 @@
+import Config
